many: move val set confdb schema and handler
