+++ asserts/confdb.go
@@ -28,203 +28,6 @@ import (
 	"github.com/snapcore/snapd/confdb"
 )
 
-var (
-	rawConfdbBuiltins = map[string]map[string][]byte{
-		"validation-sets": {
-			"headers": []byte(`type: confdb-schema
-account-id: system
-authority-id: canonical
-name: validation-sets
-summary: Manage the validation sets installed on a device
-views:
-  state:
-    summary: Read the validation sets state
-    rules:
-      -
-        request: {account}.{validation-set}
-        storage: v1.{account}.{validation-set}
-        access: read
-        content:
-          -
-            storage: snaps[{n}]
-            content:
-              -
-                storage: name
-              -
-                storage: id
-              -
-                storage: revision
-              -
-                storage: presence
-              -
-                storage: components
-                content:
-                  -
-                    storage: {component}
-                    content:
-                      -
-                        storage: presence
-                      -
-                        storage: revision
-          -
-            storage: mode
-          -
-            storage: status
-          -
-            storage: sequence
-          -
-            storage: revision
-          -
-            storage: pinned-sequence
-  admin:
-    summary: Control validation sets
-    rules:
-      -
-        request: {account}.{set-name}
-        storage: v1.{account}.{set-name}
-        access: write
-        content:
-          -
-            storage: pinned-sequence
-          -
-            storage: mode
-      -
-        request: {account}.{set-name}
-        storage: v1.{account}.{set-name}
-        access: read
-        content:
-          -
-            storage: snaps[{n}]
-            content:
-              -
-                storage: name
-              -
-                storage: id
-              -
-                storage: revision
-              -
-                storage: presence
-          -
-            storage: status
-          -
-            storage: sequence
-          -
-            storage: revision
-          -
-            storage: pinned-sequence
-          -
-            storage: mode
-  pinning-admin:
-    summary: Control pinning of validation sets
-    rules:
-      -
-        request: {account}.{set-name}.pinned-sequence
-        storage: v1.{account}.{set-name}.pinned-sequence
-        access: read-write
-`),
-			// NOTE: JSON needs to be sorted, otherwise decoding validation would fail
-			"body": []byte(`{
-  "storage": {
-    "aliases": {
-      "account-id": {
-        "pattern": "^(?:[a-z0-9A-Z]{32}|[-a-z0-9]{2,28})$",
-        "type": "string"
-      },
-      "natural-number": {
-        "min": 1,
-        "type": "int"
-      },
-      "presence": {
-        "choices": [
-          "required",
-          "optional",
-          "invalid"
-        ],
-        "type": "string"
-      },
-      "set-name": {
-        "pattern": "^[a-z0-9](?:-?[a-z0-9])*$",
-        "type": "string"
-      }
-    },
-    "schema": {
-      "v1": {
-        "keys": "${account-id}",
-        "values": {
-          "keys": "${set-name}",
-          "values": {
-            "required": [
-              "mode"
-            ],
-            "schema": {
-              "mode": {
-                "choices": [
-                  "monitor",
-                  "enforce"
-                ],
-                "type": "string"
-              },
-              "pinned-sequence": "${natural-number}",
-              "revision": "${natural-number}",
-              "sequence": "${natural-number}",
-              "snaps": {
-                "type": "array",
-                "unique": true,
-                "values": {
-                  "schema": {
-                    "components": {
-                      "values": {
-                        "schema": {
-                          "presence": "${presence}",
-                          "revision": "${natural-number}"
-                        }
-                      }
-                    },
-                    "id": "string",
-                    "name": "string",
-                    "presence": "${presence}",
-                    "revision": "${natural-number}"
-                  }
-                }
-              },
-              "status": {
-                "choices": [
-                  "valid",
-                  "invalid"
-                ],
-                "type": "string"
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-}`),
-		},
-	}
-
-	confdbSchemaCheckOrder      = []string{"type", "account-id", "authority-id"}
-	confdbSchemaExpectedHeaders = map[string]any{
-		"type":         "confdb-schema",
-		"account-id":   "system",
-		"authority-id": "canonical",
-	}
-)
-
-func init() {
-	for name, builtin := range rawConfdbBuiltins {
-		a, err := assembleBuiltinAssertion(ConfdbSchemaType, builtin["headers"], builtin["body"], builtinCheckParams{
-			order:           confdbSchemaCheckOrder,
-			expectedHeaders: confdbSchemaExpectedHeaders,
-		})
-		if err != nil {
-			panic(fmt.Sprintf("cannot create builtin %q confdb-schema: %v", name, err))
-		}
-		builtinAssertions = append(builtinAssertions, a)
-	}
-}
-
 // ConfdbSchema holds a confdb-schema assertion, which is a definition by an
 // account of access views and a storage schema for a set of related
 // configuration options under the purview of the account.
