+++ asserts/confdb_test.go
@@ -27,6 +27,7 @@ import (
 	. "gopkg.in/check.v1"
 
 	"github.com/snapcore/snapd/asserts"
+	_ "github.com/snapcore/snapd/overlord/assertstate/confdb"
 )
 
 type confdbSuite struct {
