+++ overlord/assertstate/assertstate.go
@@ -406,8 +406,6 @@ func delayedCrossMgrInit() {
 	// wire confdbstate helpers that look up confdb-schema assertions
 	confdbstate.AssertstateFetchConfdbSchemaAssertion = FetchConfdbSchemaAssertion
 	confdbstate.AssertstateConfdbSchema = ConfdbSchema
-	// register the system custodian for the "validation-sets" confdb-schema
-	confdbstate.RegisterConfdbHandler(&valsetsConfdbHandler{})
 }
 
 // AutoRefreshAssertions tries to refresh all assertions
